move all code to src/ folder
