+++ package.json
@@ -2,7 +2,7 @@
   "name": "hap-nodejs",
   "version": "0.4.53",
   "description": "HAP-NodeJS is a Node.js implementation of HomeKit Accessory Server.",
-  "main": "index.js",
+  "main": "src/index.js",
   "types": "index.d.ts",
   "engines": {
     "node": ">=4.3.2"
@@ -26,7 +26,7 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node BridgedCore.js"
+    "start": "node src/BridgedCore.js"
   },
   "repository": {
     "type": "git",
