+++ src/accessories/TemperatureSensor_accessory.js
