+++ src/lib/Accessory.js
