+++ src/lib/AccessoryLoader.js
