+++ src/lib/Bridge.js
