+++ src/lib/Camera.js
