+++ src/lib/Characteristic.js
