+++ src/lib/HAPServer.js
