+++ src/lib/StreamController.js
