+++ src/lib/gen/HomeKitTypes-Bridge.js
