+++ src/lib/model/AccessoryInfo.js
