+++ src/lib/util/clone.js
