+++ src/lib/util/encryption.js
