+++ src/lib/util/eventedhttp.js
