+++ src/lib/util/hkdf.js
